xen/memory : Add a stats_table resource type
This commit proposes a new mechanism to query the RUNSTATE_running counter for
a given vcpu from a dom0 userspace application. This commit proposes to expose
that counter by using the acquire_resource interface. For this purpose, the
commit adds a new resource named XENMEM_resource_stats_table and a
type-specific resource named XENMEM_resource_stats_table_id_vcpustats. This
type-specific resource is aiming at hosting per-VCPU counters.

The current mechanism relies on the XEN_DOMCTL_getvcpuinfo and holds a single
global domctl_lock for the entire hypercall; and iterate over every vcpu in the
system for every update thus impacting operations that share that lock.

This commit proposes to expose vcpu RUNSTATE_running via the xenforeignmemory
interface thus preventing to issue the hypercall and holding the lock. For that
purpose, a new resource type named XENMEM_resource_stats_table is added. In
particular, the type-specific resource XENMEM_resource_stats_table_id_vcpustats
is added to host per-VCPU counters. These counters are mapped in frames. The
allocation of these frames only happens if the resource is requested. The
number of allocated frames is equal to the domain's max_vcpus. Frames are
released after the domain is destroyed.

Note that the updating of vcpu's counters is in a hot path, thus, in this commit,
copying only happens if it is specifically required.

Note that the resource is extensible in two ways. First, the structure
vcpu_stats can be extended with new per-vcpu counters. To do so, new per-vcpu
counters shall be added after the last element of the structure definition to
not break existing consumers. Second, new type-specific resources can be added
to host a different sort of counters.

Signed-off-by: Matias Ezequiel Vara Larsen <matias.vara@vates.fr>