x86/mm: Do not validate/devalidate PGT_none type
This commit prevents PGT_none type pages to be validated/devalidated. This is
required for the use-case in which a guest-agnostic resource is allocated. In
this case, these pages may be accessible by an "owning" PV domain. To lock the
page from guest pov, pages are required to be marked with PGT_none with a
single type ref. This commit prevents triggering the BUGs() at
validate_page()/devalidate_page().

Signed-off-by: Matias Ezequiel Vara Larsen <matias.vara@vates.fr>